+++ EZOCore.lua
@@ -2,8 +2,8 @@ EZOCore = EZOCore or {}
 local EZOCore = EZOCore
 
 EZOCore.name = "EZOCore"
-EZOCore.version = "0.1.1"
-EZOCore.addOnVersion = 101
+EZOCore.version = "0.1.2"
+EZOCore.addOnVersion = 102
 EZOCore.apiVersion = 1
 
 EZOCore.EVENT_INITIALIZED = "EZO_CORE_INITIALIZED"